server/requirements.txt
# Google Agent Development Kit and AI Core
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.39.0
litellm>=1.0.0

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0

# Data processing and caching
pandas>=2.0.0
numpy>=1.24.0
redis>=5.0.0

# Database and persistence
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Date and time handling
python-dateutil>=2.8.0

# Async support
aiofiles>=23.0.0
aiohttp>=3.9.0

# Security and authentication
cryptography>=41.0.0
PyJWT>=2.8.0

# Optimization and ML
scikit-learn>=1.3.0
joblib>=1.3.0

# API integrations
instacart-python>=1.0.0  # Custom Instacart integration
requests>=2.31.0
